+++ test/scripts/deposit.js
@@ -2,7 +2,7 @@ const BatchExchange = artifacts.require("BatchExchange")
 const GnosisSafe = artifacts.require("GnosisSafe")
 const ProxyFactory = artifacts.require("GnosisSafeProxyFactory")
 
-const { addCustomMintableTokenToExchange, deploySafe } = require("../../scripts/utils/process_flows")(web3, artifacts)
+const { addCustomMintableTokenToExchange, deploySafe } = require("../../scripts/utils/strategy_simulator")(web3, artifacts)
 const { deployFleetOfSafes, buildDepositFromList } = require("../../scripts/utils/trading_strategy_helpers")(web3, artifacts)
 const { waitForNSeconds, execTransaction } = require("../../scripts/utils/internals")(web3, artifacts)
